注册登录

+++ src/main/java/com/kk/pojo/Reader.java
@@ -2,10 +2,12 @@ package com.kk.pojo;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class Reader {
     private int id;
-    private int sno;
+    private BigInteger sno;
     private String pwd;
     private String name;
     private String major;
